+++ prp/parse/spatyper.py
@@ -10,10 +10,11 @@ from ..models.typing import TypingSoftware as Software
 LOG = logging.getLogger(__name__)
 
 
-def parse_spatyper_results(path:str) -> SpatyperTypingMethodIndex:
+def parse_spatyper_results(path: str) -> SpatyperTypingMethodIndex:
     """Read spaTyper output file."""
     LOG.info("Parsing spatyper results")
-    result = (
+
+    result_loa = (
         pd.read_csv(path, delimiter="\t")
         .rename(
             columns={
@@ -22,23 +23,19 @@ def parse_spatyper_results(path:str) -> SpatyperTypingMethodIndex:
                 "Type": "type"
             }
         )
+        .to_dict(orient="records")
     )
-    # extract the values from the first row (index 0)
-    first_row = result.iloc[0].str.strip()
-    
-    # create typing result object
+
+    result = result_loa[0] if result_loa else {}
+
     result_obj = TypingResultSpatyper(
-        sequence_name = first_row["sequence_name"],
-        repeats = first_row["repeats"],
-        type = first_row["type"]
+        sequence_name=result.get("sequence_name"),
+        repeats=result.get("repeats"),
+        type=result.get("type")
     )
+
     return SpatyperTypingMethodIndex(
-    type=TypingMethod.SPATYPE, software=Software.SPATYPER, result=result_obj
+        type=TypingMethod.SPATYPE,
+        software=Software.SPATYPER,
+        result=result_obj
     )
-
-     # How does it look if nothing is found? Account for missing result? See below.
-
-     # Sequence name	Repeats	Type
-     # nothing in this row
-
-     # do I have to explicitly say that results can be None?